update  the docker file
this non escaped slash caused a failure to build. Other errors not found but the instructions could be a little cleaner around stopping old containers. I suspect the issue is with inspecting a running instance and a failure to build the new instance and inspect that one OR this could be environment specific.